requirements.txt
Flask_WTF==0.14.3
Flask_SocketIO==5.0.1
Flask==1.1.2
Flask_Login==0.5.0
SQLAlchemy_serializer==1.3.4.4
Werkzeug==1.0.1
SQLAlchemy==1.3.23
WTForms==2.3.3
Flask_RESTful==0.3.8
Pillow==8.2.0
